LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/uievents/mouse/mousemove_prevent_default_action.tentative-expected.txt

div a
div b
div c

FAIL selectionchange event firing when mousemove event is prevented assert_equals: received events expected "mousemove,mousedown,selectionchange,mousemove,selectionchange" but got "mousemove,mousedown,selectionchange,mousemove,mousemove"
FAIL dragstart event firing when mousemove event is prevented assert_equals: received events expected "mousemove,mousedown,mousemove,dragstart" but got "mousemove,mousedown,selectionchange,mousemove,mousemove"
